Moved basic bigtuna config before rails is loaded
